Convert test_verify unit tests to pytests
